fix: adjust feedback
