--- packages/map/src/overlay/bubble/index.ts
@@ -1,2 +1,0 @@
-export * from './bubble-base'
-export * from './wrapper'
