--- packages/map/src/overlay/bubble/wrapper/index.ts
@@ -1,2 +1,0 @@
-export * from './heart'
-export * from './link'
