--- packages/map/src/overlay/circle-marker/index.ts
@@ -1,9 +1,0 @@
-import withTypeCircleMarker from './with-type-circle-marker'
-
-export * from './circle-marker'
-
-/** 공통으로 사용되는 poi 마커 */
-export const HotelCircleMarker = withTypeCircleMarker('hotel')
-export const AttractionCircleMarker = withTypeCircleMarker('attraction')
-export const RestaurantCircleMarker = withTypeCircleMarker('restaurant')
-export const TnaCircleMarker = withTypeCircleMarker('tna')
