+++ packages/map/src/overlay/index.ts
@@ -1,4 +1,3 @@
 export * from './polyline'
 export * from './polygon'
-export * from './circle-marker'
-export * from './bubble'
+export * from './markers'
