+++ packages/map/src/overlay/markers/index.ts
@@ -1,0 +1,2 @@
+export * from './primary-marker'
+export * from './poi-dot-marker'
